+++ CHANGES.md
@@ -1,1 +1,2 @@
 - Add filter setting for max cost per artisan material bag to NPC orders page
+- Add buying automation button to Auctionator shopping tab